create function readArticleSubgenre.js to retrieve subgenre data
